+++ sparkles-core/src/headers.rs
@@ -6,20 +6,22 @@ use crate::local_storage::id_mapping::IdStoreMap;
 #[derive(Serialize, Deserialize, Clone, Debug, Default)]
 pub struct LocalPacketHeader {
     pub thread_ord_id: u64,
-    pub thread_id: u64,
+    pub thread_info: Option<ThreadInfo>,
 
     pub start_timestamp: u64,
     pub end_timestamp: u64,
 
     pub id_store: IdStoreMap,
-
-    // Keep it here to make parser simpler
-    pub counts_per_ns: f64,
 }
 
-/// This header describe updated thread name
+#[derive(Serialize, Deserialize, Clone, Debug, Default)]
+pub struct ThreadInfo {
+    pub thread_id: u64,
+    pub new_thread_name: Option<String>,
+}
+
 #[derive(Serialize, Deserialize, Clone, Debug)]
-pub struct ThreadNameHeader {
-    pub thread_ord_id: u64,
-    pub thread_name: String,
+pub struct SparklesEncoderInfo {
+    pub ver: u32,
+    pub counts_per_ns: f64,
 }