+++ sparkles-core/src/lib.rs
@@ -6,3 +6,6 @@ pub use timestamp::{Timestamp, TimestampProvider};
 
 pub mod local_storage;
 pub mod headers;
+pub mod config;
+
+pub const ENCODER_VER: u32 = 0;