+++ sparkles/examples/how_to_use.rs
@@ -7,14 +7,13 @@ use std::thread;
 use std::time::Duration;
 use log::LevelFilter;
 use simple_logger::SimpleLogger;
-use sparkles::SparklesConfigBuilder;
-use sparkles_macro::{instant_event, range_event_start};
+use sparkles_macro::{instant_event, range_event_end, range_event_start};
 
 fn main() {
     SimpleLogger::default().with_level(LevelFilter::Debug).init().unwrap();
     // Init and acquire finalize guard to automatically finalize event collection and 
     // flush them to the destination when the main thread finished
-    let finalize_guard = SparklesConfigBuilder::default_init();
+    let finalize_guard = sparkles::init_default();
     // Start range event
     // It's finished when guard is dropped
     let g = range_event_start!("main()");
@@ -35,6 +34,7 @@ fn main() {
             instant_event!("*_*");
             thread::sleep(Duration::from_micros(1_000));
         }
+        range_event_end!(g, "custom range end message!");
         // sparkles::flush_thread_local();
     }).unwrap();