+++ sparkles/src/thread_local_storage.rs
@@ -1,22 +1,21 @@
 use std::cell::RefCell;
+use std::sync::OnceLock;
 use std::thread;
-use sparkles_core::headers::{LocalPacketHeader, ThreadNameHeader};
+use sparkles_core::config::LocalStorageConfig;
+use sparkles_core::headers::{LocalPacketHeader, ThreadInfo};
 use sparkles_core::local_storage::{GlobalStorageImpl, LocalStorage};
-use crate::global_storage::GLOBAL_STORAGE;
+use crate::global_storage::{GlobalStorage, GLOBAL_STORAGE};
 
 pub struct GlobalStorageRef;
 pub type ThreadLocalStorage = LocalStorage<GlobalStorageRef>;
 
+static LOCAL_CONFIG: OnceLock<LocalStorageConfig> = OnceLock::new();
+
 impl GlobalStorageImpl for GlobalStorageRef {
-    fn flush(&self, header: LocalPacketHeader, data: Vec<u8>) {
+    fn flush(&self, header: &LocalPacketHeader, data: Vec<u8>) {
         let mut global_storage_ref = GLOBAL_STORAGE.lock().unwrap();
-        let global_storage_ref = global_storage_ref.get_or_insert_with(Default::default);
+        let global_storage_ref = global_storage_ref.get_or_insert_with(|| GlobalStorage::new(Default::default()));
         global_storage_ref.push_buf(header, &data);
-    }
-    fn put_thread_name(&self, header: ThreadNameHeader) {
-        let mut global_storage_ref = GLOBAL_STORAGE.lock().unwrap();
-        let global_storage_ref = global_storage_ref.get_or_insert_with(Default::default);
-        global_storage_ref.update_thread_name(header);
     }
 }
 
@@ -24,7 +23,12 @@ fn new_local_storage() -> LocalStorage<GlobalStorageRef> {
     let thread_info = thread::current();
     let thread_name = thread_info.name().unwrap_or("Unnamed thread").to_string();
     let thread_id = thread_id::get() as u64;
-    LocalStorage::new(GlobalStorageRef, thread_name, thread_id)
+    let thread_info = ThreadInfo {
+        new_thread_name: Some(thread_name.clone()),
+        thread_id,
+    };
+    let config = *LOCAL_CONFIG.get_or_init(LocalStorageConfig::default);
+    LocalStorage::new(GlobalStorageRef, Some(thread_info), config)
 }
 
 #[inline(always)]